bug fixes; update python path; update doc

+++ bsubScripts/ncum_global_hycom_input/hycom_create_tarball_g2files_put_into_ftp.py
@@ -23,14 +23,14 @@ def createTarBalls(path, today, utc, stephr=3):
     if utc != '00': return
     cdir = os.getcwd()
     os.chdir(path)
-    anal_ftemp = 'anal*%s*%s.grb2'
+    anal_ftemp = 'ncum_anal*%s*%s.grb2'
     tDay = datetime.datetime.strptime(today, "%Y%m%d")
     lag = datetime.timedelta(days=1)
     yDay = (tDay - lag).strftime('%Y%m%d')
     # get yesterday's analysis files from 06hr onwards
-    yanal_files = [anal_ftemp % (str(hr).zfill(2), yDay) for hr in range(6, 24, stephr)]
+    yanal_files = [anal_ftemp % (yDay, str(hr).zfill(2)) for hr in range(6, 24, stephr)]
     # get today's analysis 00 and 03 hr
-    tanal_files = [anal_ftemp % (str(hr).zfill(2), today) for hr in range(0, 6, stephr)]
+    tanal_files = [anal_ftemp % (today, str(hr).zfill(2)) for hr in range(0, 6, stephr)]
     for yf in yanal_files:
         # move yesterday's analysis files to today's directory
         cmd = 'mv ../%s/%s .' % (yDay, yf)
@@ -45,33 +45,43 @@ def createTarBalls(path, today, utc, stephr=3):
     # create analysis files tar file in parallel
     anal_files = '  '.join(yanal_files + tanal_files)  
     cmd = "tar -c  %s | %s -v  -c -f -p32 -m500 > %s/ncum_anal_%s.tar.bz2" % (anal_files, pbzip2, '../TarFiles', today)
+    print cmd
     subprocess.call(cmd, shell=True)
     # create forecast files tar file in parallel
-    cmd = "tar -c fcst*.grb2 | %s  -v  -c -f -p32 -m500 > %s/ncum_fcst_%s.tar.bz2" % (pbzip2, '../TarFiles', today)
+    cmd = "tar -c ncum_fcst*.grb2 | %s  -v  -c -f -p32 -m500 > %s/ncum_fcst_%s.tar.bz2" % (pbzip2, '../TarFiles', today)
+    print cmd
     subprocess.call(cmd, shell=True)
     
     # delete analysis & forecasts files, after tar ball has been created!
-    cmd = "rm %s" % anal_files
+    cmd = "rm -rf %s" % anal_files
+    print cmd
     subprocess.call(cmd, shell=True)
-    cmd = "rm ncum_fcst_%s*.grb2" % today
+    cmd = "rm -rf ncum_fcst_%s*.grb2" % today
+    print cmd
     subprocess.call(cmd, shell=True)
     
     # remove yesterday's empty directory, not today directory!!!
     yDayPath = os.path.join(path, '../%s' % yDay)
-    os.rmdir(yDayPath)  
+    print cmd
     
-    # do scp the tar files to ftp_server and nkn_server
-    cmd = 'rsh ncmr0102 "scp -p ../TarFiles/ncum_anal_%s.tar.bz2  %s:/data/ftp/pub/outgoing/NCUM_INCOIS/Hycom/"' % (today, ftp_server)
-    subprocess.call(cmd, shell=True)
-    cmd = 'rsh ncmr0102 "scp -p ../TarFiles/ncum_anal_%s.tar.bz2  %s:NCUM/hycom/"' % (today, nkn_server)
-    subprocess.call(cmd, shell=True)
-    
-    cmd = 'rsh ncmr0102 "scp -p ../TarFiles/ncum_fcst_%s.tar.bz2  %s:/data/ftp/pub/outgoing/NCUM_INCOIS/Hycom/"' % (today, ftp_server)
-    subprocess.call(cmd, shell=True)
-    cmd = 'rsh ncmr0102 "scp -p ../TarFiles/ncum_fcst_%s.tar.bz2  %s:NCUM/hycom/"' % (today, nkn_server)
-    subprocess.call(cmd, shell=True)
-    
+    if not os.listdir(yDayPath): os.rmdir(yDayPath)    
     os.chdir(cdir)  
+    
+    tarpath = os.path.abspath('../TarFiles')
+    # do scp the tar files to ftp_server and nkn_server
+    cmd = 'ssh ncmlogin3 "scp -p %s/ncum_anal_%s.tar.bz2  %s:/data/ftp/pub/outgoing/NCUM_INCOIS/Hycom/"' % (tarpath, today, ftp_server)
+    print cmd
+    subprocess.call(cmd, shell=True)
+    cmd = 'ssh ncmlogin3 "scp -p %s/ncum_anal_%s.tar.bz2  %s:NCUM/hycom/"' % (tarpath, today, nkn_server)
+    print cmd
+    subprocess.call(cmd, shell=True)
+    
+    cmd = 'ssh ncmlogin3 "scp -p %s/ncum_fcst_%s.tar.bz2  %s:/data/ftp/pub/outgoing/NCUM_INCOIS/Hycom/"' % (tarpath, today, ftp_server)
+    print cmd
+    subprocess.call(cmd, shell=True)
+    cmd = 'ssh ncmlogin3 "scp -p %s/ncum_fcst_%s.tar.bz2  %s:NCUM/hycom/"' % (tarpath, today, nkn_server)
+    print cmd
+    subprocess.call(cmd, shell=True)
 # end of def createTarBalls(path, today, ...):
 
 if __name__ == '__main__':
